Refactoring, deleted un-used classes since they have now been split up into use cases, controllers, or simply just do not need them anymore
