+++ src/main/java/CalculatorMain.java
@@ -1,4 +1,6 @@
+import Controller.CalculatorController;
 import Entity.ExpressionException;
+import UI.CalculatorUI;
 import UseCase.BasicOperations.ParserException;
 
 import java.io.IOException;
@@ -8,7 +10,10 @@ public class CalculatorMain {
     public static void main(String[] args) throws IOException, ClassNotFoundException, ParserException, ExpressionException {
 
         System.out.println("Welcome to The Basis Calculator!");
-        UI.CalculatorUI.main();
+        CalculatorController control = new CalculatorController();
+        CalculatorUI new_ui = new CalculatorUI();
+        new_ui.setControl(control);
+        new_ui.main();
     }
 
 }
