+++ src/main/java/Frontend/MyController.java
@@ -1,7 +1,5 @@
 package Frontend;
 
-import UseCase.InputProcessor;
-
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
