+++ src/main/java/UseCase/Accounts/UserLogIn.java
@@ -1,6 +1,6 @@
 package UseCase.Accounts;
 
-import Entity.UserEntity;
+import Entity.User;
 import Gateway.UserList;
 
 import java.io.IOException;
@@ -8,11 +8,11 @@ import java.util.*;
 
 public class UserLogIn implements LogIn{
 
-    private final UserEntity user = new UserEntity();
+    private final User user = new User();
     private final UserList list = new UserList();
 
     @Override
-    public UserEntity logIn() throws IOException, ClassNotFoundException {
+    public User logIn() throws IOException, ClassNotFoundException {
         ValidateLogin validateLogin = new ValidateLogin(user, list);
 
         Scanner info = new Scanner(System.in);
